Split upscale recent jobs hook
